Refactoring structures to Windows.ShellScalingApi project
