+++ src/PInvoke.sln
@@ -244,14 +244,29 @@ Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "SHCore.Shared", "SHCore.Sha
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SHCore.Tests", "SHCore.Tests\SHCore.Tests.csproj", "{49094447-F41D-4026-92E3-124532958876}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Windows.ShellScalingApi.Desktop", "Windows.ShellScalingApi.Desktop\Windows.ShellScalingApi.Desktop.csproj", "{0500E681-818F-4366-B2A5-0BB73D7864C6}"
+EndProject
+Project("{FF286327-C783-4F7A-AB73-9BCBAD0D4460}") = "Windows.ShellScalingApi.NuGet", "Windows.ShellScalingApi.NuGet\Windows.ShellScalingApi.NuGet.nuproj", "{CB9CFE56-EA10-412C-BEF9-1FA7C38AA76E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Windows.ShellScalingApi.Profile111", "Windows.ShellScalingApi.Profile111\Windows.ShellScalingApi.Profile111.csproj", "{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}"
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Windows.ShellScalingApi.Shared", "Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.shproj", "{076AE6A9-D318-47C7-A1E6-9F8022EEC7E2}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Windows.ShellScalingApi.Tests", "Windows.ShellScalingApi.Tests\Windows.ShellScalingApi.Tests.csproj", "{B095EF1F-BE56-47BC-91AF-510F0531EC1E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Windows.ShellScalingApi", "Windows.ShellScalingApi\Windows.ShellScalingApi.csproj", "{A8669E42-8247-448D-80A8-5575ED7C63E6}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		Msi.Shared\Msi.Shared.projitems*{01174e89-de65-47dc-bf53-4e2cf34ad843}*SharedItemsImports = 4
+		Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems*{0500e681-818f-4366-b2a5-0bb73d7864c6}*SharedItemsImports = 4
 		CodeGenerationAttributes.Shared\CodeGenerationAttributes.Shared.projitems*{05fba24a-b0f6-4b04-b00f-9aa5acd191a9}*SharedItemsImports = 4
+		Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems*{076ae6a9-d318-47c7-a1e6-9f8022eec7e2}*SharedItemsImports = 13
 		NCrypt.Shared\NCrypt.Shared.projitems*{093be65b-820a-470c-9a47-82592400059f}*SharedItemsImports = 4
 		Windows.Core.Shared\Windows.Core.Shared.projitems*{1386982c-4731-46cd-9444-18e891aecde0}*SharedItemsImports = 4
 		CodeGenerationAttributes.Shared\CodeGenerationAttributes.Shared.projitems*{1387e009-7086-4572-ac8d-ce4242adec77}*SharedItemsImports = 4
 		Kernel32.Shared\Kernel32.Shared.projitems*{15a8ded1-0f36-4b35-9973-22851da77544}*SharedItemsImports = 13
+		Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems*{1878e4ac-cd1a-4a5b-817a-af6fa9c0020d}*SharedItemsImports = 4
 		Kernel32.Shared\Kernel32.Shared.projitems*{18ba8c60-9a63-4eea-bb5c-c9899d86f5b8}*SharedItemsImports = 4
 		Kernel32.Shared\Kernel32.Shared.projitems*{1fdc092a-d3e8-4cc4-b896-17e0a213f723}*SharedItemsImports = 4
 		Crypt32.Shared\Crypt32.Shared.projitems*{2355a580-2257-4a41-b7ec-f987993ce0c9}*SharedItemsImports = 4
@@ -280,6 +295,7 @@ Global
 		NetApi32.Shared\NetApi32.Shared.projitems*{9c5ae8e9-330a-47d9-84d8-9883880ebe9e}*SharedItemsImports = 4
 		BCrypt.Shared\BCrypt.Shared.projitems*{9ce056e3-5d53-4092-97cc-78826a4d1c14}*SharedItemsImports = 4
 		DbgHelp.Shared\DbgHelp.Shared.projitems*{9e993e23-3da3-4824-afd1-d8c9db1d9c8a}*SharedItemsImports = 13
+		Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems*{a8669e42-8247-448d-80a8-5575ed7c63e6}*SharedItemsImports = 4
 		NTDll.Shared\NTDll.Shared.projitems*{a8eafcaf-fcd3-4255-93fa-1a51ca9aa07c}*SharedItemsImports = 4
 		NCrypt.Shared\NCrypt.Shared.projitems*{a9b393ca-9f5f-41a7-832c-26b92091b2f7}*SharedItemsImports = 4
 		Windows.Core.Shared\Windows.Core.Shared.projitems*{b08c3c79-4cdd-4d37-933c-07d3452fd5f1}*SharedItemsImports = 4
@@ -676,6 +692,26 @@ Global
 		{49094447-F41D-4026-92E3-124532958876}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{49094447-F41D-4026-92E3-124532958876}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{49094447-F41D-4026-92E3-124532958876}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0500E681-818F-4366-B2A5-0BB73D7864C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0500E681-818F-4366-B2A5-0BB73D7864C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0500E681-818F-4366-B2A5-0BB73D7864C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0500E681-818F-4366-B2A5-0BB73D7864C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CB9CFE56-EA10-412C-BEF9-1FA7C38AA76E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CB9CFE56-EA10-412C-BEF9-1FA7C38AA76E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CB9CFE56-EA10-412C-BEF9-1FA7C38AA76E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CB9CFE56-EA10-412C-BEF9-1FA7C38AA76E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B095EF1F-BE56-47BC-91AF-510F0531EC1E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B095EF1F-BE56-47BC-91AF-510F0531EC1E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B095EF1F-BE56-47BC-91AF-510F0531EC1E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B095EF1F-BE56-47BC-91AF-510F0531EC1E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8669E42-8247-448D-80A8-5575ED7C63E6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8669E42-8247-448D-80A8-5575ED7C63E6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8669E42-8247-448D-80A8-5575ED7C63E6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8669E42-8247-448D-80A8-5575ED7C63E6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
