+++ src/User32.Desktop/User32.Desktop.csproj
@@ -23,7 +23,6 @@
   <ItemGroup>
     <Compile Include="User32+InputType.cs" />
     <Compile Include="User32+KEYEVENTF.cs" />
-    <Compile Include="User32+PROCESS_DPI_AWARENESS.cs" />
     <Compile Include="User32+MOUSEEVENTF.cs" />
     <Compile Include="User32+MOUSEINPUT.cs">
       <Generator>MSBuild:GenerateCodeFromAttributes</Generator>
@@ -84,6 +83,10 @@
     <ProjectReference Include="..\Windows.Core\Windows.Core.csproj">
       <Project>{b08c3c79-4cdd-4d37-933c-07d3452fd5f1}</Project>
       <Name>Windows.Core</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Windows.ShellScalingApi.Desktop\Windows.ShellScalingApi.Desktop.csproj">
+      <Project>{0500e681-818f-4366-b2a5-0bb73d7864c6}</Project>
+      <Name>Windows.ShellScalingApi.Desktop</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
