+++ src/User32.NuGet/User32.NuGet.nuproj
@@ -36,6 +36,7 @@
   <ItemGroup>
     <ProjectReference Include="..\User32.Desktop\User32.Desktop.csproj" />
     <ProjectReference Include="..\Windows.Core.NuGet\Windows.Core.NuGet.nuproj" />
+    <ProjectReference Include="..\Windows.ShellScalingApi.NuGet\Windows.ShellScalingApi.NuGet.nuproj" />
   </ItemGroup>
   <ItemGroup>
     <None Include="PInvoke.User32.nuspec" />
