+++ src/Windows.ShellScalingApi.Desktop/ShellScallingApi+SHELL_UI_COMPONENT.cs
