+++ src/Windows.ShellScalingApi.Desktop/project.json
@@ -1,0 +1,11 @@
+﻿{
+  "supports": { },
+  "dependencies": {
+  },
+  "frameworks": {
+    ".NETFramework,Version=v4.0": { }
+  },
+  "runtimes": {
+    "win": { }
+  }
+}
