+++ src/Windows.ShellScalingApi.Profile111/Properties/AssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+
+using System.Reflection;
+
+[assembly: AssemblyTitle("PInvoke.Windows.ShellScalingApi")]
+[assembly: AssemblyProduct("PInvoke")]
+[assembly: AssemblyCopyright("Copyright © Andrew Arnott 2015")]
