+++ src/Windows.ShellScalingApi.Profile111/Windows.ShellScalingApi.Profile111.csproj
@@ -9,11 +9,11 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{9C08E127-1B44-4738-9C6C-89244A3A6B53}</ProjectGuid>
+    <ProjectGuid>{1878E4AC-CD1A-4A5B-817A-AF6FA9C0020D}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>PInvoke</RootNamespace>
-    <AssemblyName>PInvoke.SHCore</AssemblyName>
+    <AssemblyName>PInvoke.Windows.ShellScalingApi</AssemblyName>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <MinimumVisualStudioVersion>10.0</MinimumVisualStudioVersion>
@@ -45,12 +45,8 @@
       <Project>{b08c3c79-4cdd-4d37-933c-07d3452fd5f1}</Project>
       <Name>Windows.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Windows.ShellScalingApi.Profile111\Windows.ShellScalingApi.Profile111.csproj">
-      <Project>{1878e4ac-cd1a-4a5b-817a-af6fa9c0020d}</Project>
-      <Name>Windows.ShellScalingApi.Profile111</Name>
-    </ProjectReference>
   </ItemGroup>
-  <Import Project="..\SHCore.Shared\SHCore.Shared.projitems" Label="Shared" />
+  <Import Project="..\Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))\EnlistmentInfo.targets" Condition=" '$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))' != '' " />
 </Project>
