+++ src/Windows.ShellScalingApi.Profile111/project.json
@@ -1,0 +1,8 @@
+﻿{
+  "supports": { },
+  "dependencies": {
+  },
+  "frameworks": {
+    ".NETPortable,Version=v4.5,Profile=Profile111": { }
+  }
+}
