+++ src/Windows.ShellScalingApi.Tests/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+
+using System.Reflection;
+using System.Resources;
+
+[assembly: AssemblyTitle("PInvoke.Windows.ShellScalingApi.Tests")]
+[assembly: AssemblyProduct("PInvoke.Windows.ShellScalingApi.Tests")]
+[assembly: AssemblyCopyright("Copyright ©  2015")]
+[assembly: NeutralResourcesLanguage("en")]
