+++ src/Windows.ShellScalingApi.Tests/Windows.ShellScalingApiFacts.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+
+using System;
+using PInvoke;
+using Xunit;
+using static PInvoke.Windows.ShellScalingApi;
+
+public class Windows.ShellScalingApiFacts
+{
+    [Fact(Skip = "No tests yet")]
+    public void NoTests()
+    {
+    }
+}
