+++ src/Windows.ShellScalingApi/Windows.ShellScalingApi.csproj
@@ -1,19 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="14.0">
   <PropertyGroup>
-    <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
-    <TargetFrameworkProfile>Profile111</TargetFrameworkProfile>
+    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
+    <TargetFrameworkProfile>Profile92</TargetFrameworkProfile>
   </PropertyGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.props))\EnlistmentInfo.props" Condition=" '$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.props))' != '' " />
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{9C08E127-1B44-4738-9C6C-89244A3A6B53}</ProjectGuid>
+    <ProjectGuid>{A8669E42-8247-448D-80A8-5575ED7C63E6}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>PInvoke</RootNamespace>
-    <AssemblyName>PInvoke.SHCore</AssemblyName>
+    <AssemblyName>PInvoke.Windows.ShellScalingApi</AssemblyName>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
     <MinimumVisualStudioVersion>10.0</MinimumVisualStudioVersion>
@@ -29,9 +29,9 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\CodeGenerationAttributes.Profile111\CodeGenerationAttributes.Profile111.csproj">
-      <Project>{05FBA24A-B0F6-4B04-B00F-9AA5ACD191A9}</Project>
-      <Name>CodeGenerationAttributes.Profile111</Name>
+    <ProjectReference Include="..\CodeGenerationAttributes\CodeGenerationAttributes.csproj">
+      <Project>{1387e009-7086-4572-ac8d-ce4242adec77}</Project>
+      <Name>CodeGenerationAttributes</Name>
       <Private>False</Private>
       <ExcludeFromNuPkg>true</ExcludeFromNuPkg>
     </ProjectReference>
@@ -45,12 +45,8 @@
       <Project>{b08c3c79-4cdd-4d37-933c-07d3452fd5f1}</Project>
       <Name>Windows.Core</Name>
     </ProjectReference>
-    <ProjectReference Include="..\Windows.ShellScalingApi.Profile111\Windows.ShellScalingApi.Profile111.csproj">
-      <Project>{1878e4ac-cd1a-4a5b-817a-af6fa9c0020d}</Project>
-      <Name>Windows.ShellScalingApi.Profile111</Name>
-    </ProjectReference>
   </ItemGroup>
-  <Import Project="..\SHCore.Shared\SHCore.Shared.projitems" Label="Shared" />
+  <Import Project="..\Windows.ShellScalingApi.Shared\Windows.ShellScalingApi.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))\EnlistmentInfo.targets" Condition=" '$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildProjectDirectory), EnlistmentInfo.targets))' != '' " />
 </Project>
