+++ src/Windows.ShellScalingApi/project.json
@@ -1,0 +1,8 @@
+﻿{
+  "supports": { },
+  "dependencies": {
+  },
+  "frameworks": {
+    ".NETPortable,Version=v4.0,Profile=Profile92": { }
+  }
+}
